+++ akshare/__init__.py
@@ -2167,9 +2167,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.7.47 add: add stock_zh_valuation_baidu interface
 1.7.48 add: add stock_zh_vote_baidu interface
 1.7.49 add: add futures_news_baidu interface
+1.7.50 add: add stock_hot_search_baidu interface
 """
 
-__version__ = "1.7.49"
+__version__ = "1.7.50"
 __author__ = "AKFamily"
 
 import sys
@@ -2179,6 +2180,11 @@ if sys.version_info < (3, 7):
     sys.exit(1)
 
 del sys
+
+"""
+百度股市通-热搜股票
+"""
+from akshare.stock.stock_hot_search_baidu import stock_hot_search_baidu
 
 """
 百度股市通-期货-新闻